more moves
- mv selector and matcher to select mod
- separate parsing out, and move it to query mod (with a trivial
  delegation from select)
- rename select::Selector trait, so that it doesn't conflict with
  query::Selector
