+++ src/select/sel_table.rs
@@ -1,6 +1,6 @@
 use crate::matcher::StringMatcher;
 use crate::query::TableSliceMatcher;
-use crate::select::Selector;
+use crate::select::TrySelector;
 use crate::tree_ref::{MdElemRef, TableSlice};
 
 #[derive(Debug, PartialEq)]
@@ -9,7 +9,7 @@ pub struct TableSliceSelector {
     rows_matcher: StringMatcher,
 }
 
-impl<'md> Selector<'md, TableSlice<'md>> for TableSliceSelector {
+impl<'md> TrySelector<'md, TableSlice<'md>> for TableSliceSelector {
     fn try_select(&self, slice: TableSlice<'md>) -> Option<MdElemRef<'md>> {
         let mut slice = slice.clone(); // GH #168 is there any way to avoid this? There may not be.
         slice.normalize();
@@ -40,7 +40,7 @@ impl From<TableSliceMatcher> for TableSliceSelector {
 mod tests {
     use super::*;
 
-    use crate::select::Selector;
+    use crate::select::TrySelector;
     use crate::tree::{Inline, Line, Table, Text, TextVariant};
     use crate::utils_for_test::*;
     use markdown::mdast;
